Tweak web crawler restrictions, Index and Query using whoosh

+++ Assignment 4/main.py
@@ -10,13 +10,18 @@ import sys
 IND_PATH = '_index'
 DOCS_PATH = 'docs'
 URL_MAP_PATH = '_url_map.dat'
+REBUILD = False
 
 def create_index(fn):
+    schema = Schema(url=ID(stored=True, unique=True), title=TEXT(stored=True), content=TEXT(stored=True))
+    
     if os.path.exists(fn):
-        shutil.rmtree(fn)
+        if REBUILD:
+            shutil.rmtree(fn)
+        else:
+            return index.open_dir(fn, schema=schema)
 
     os.makedirs(fn)
-    schema = Schema(url=ID(stored=True, unique=True), title=TEXT(stored=True), content=TEXT(stored=True))
     return index.create_in(fn, schema)
 
 def unpickle_url_map(fn):
@@ -33,8 +38,8 @@ def unpickle_url_map(fn):
 
 def extract_content(html):
     page = BeautifulSoup(html, 'html.parser')
-    title = page.title.string if page.title else 'No Title'
-    content = page.get_text().lower()
+    title = str(page.title.string) if page.title else 'No Title'
+    content = page.get_text(strip=True).lower()
     return title, content
 
 def add_docs(ind, docs_fn, url_map):
@@ -43,17 +48,13 @@ def add_docs(ind, docs_fn, url_map):
         sys.exit(1)
 
     writer = ind.writer()
-    cnt = 1
-    batch_size = 2
 
     for url, fn in url_map.items():
         if os.path.isfile(fn):
             with open(fn, 'r', encoding='utf-8') as file:
                 title, content = extract_content(file.read())
                 writer.add_document(url=url, title=title, content=content)
-                if cnt % batch_size == 0:
-                    writer.commit()
-                    writer = ind.writer()
+                print("added document")
     
     writer.commit()
 
@@ -66,17 +67,21 @@ def query_session(ind):
                 query = QueryParser("content", ind.schema).parse(user_query)
                 results = searcher.search(query, terms=True)
                 
-                for r in results:
-                    print(r, r.score)
-                    if results.has_matched_terms():
-                        print(results.matched_terms())
+                for hit in results:
+                    print('--------------------------------------------------------------------------')
+                    print('Title:', hit['title'])
+                    print('Score:\n  ', hit.score)
+                    print('Highlights:\n')
+                    print(hit.highlights("content"))
+                    print()
     except KeyboardInterrupt:
         print('Exiting...')
 
 def main():
     ind = create_index(IND_PATH)
     url_map = unpickle_url_map(URL_MAP_PATH)
-    add_docs(ind, DOCS_PATH, url_map)
+    if REBUILD:
+        add_docs(ind, DOCS_PATH, url_map)
     query_session(ind)
 
 if __name__ == '__main__':
